Move receipt payloads from validator_requests.py to schemas.py
