+++ compute_horde/compute_horde/receipts/models.py
@@ -3,11 +3,11 @@ from datetime import timedelta
 from django.db import models
 
 from compute_horde.executor_class import DEFAULT_EXECUTOR_CLASS, ExecutorClass
-from compute_horde.mv_protocol.validator_requests import (
+from compute_horde.receipts.schemas import (
     JobFinishedReceiptPayload,
     JobStartedReceiptPayload,
+    Receipt,
 )
-from compute_horde.receipts.schemas import Receipt
 
 
 class ReceiptNotSigned(Exception):
